Bug 816365: Changes to sahpi.c and hpifru.c to suppress inconsistent type warnings in gcc-3.3-23. Casted int to unsigned int when reasonable or changed int declaration to unsigned int, or ((int)i < (uint)u) to ((int)i < 0 || (uint)i < (uint)u). Second opinions are welcome, please.

+++ util/hpifru.c
@@ -61,7 +61,7 @@ prtchassinfo(void)
 {
   chasstype = (SaHpiInventChassisTypeT)inv->DataRecords[i]->RecordData.ChassisInfo.Type;
   for (k=0; k<NCT; k++) {
-	  if (k == chasstype)
+	  if ((unsigned int)k == chasstype)
 		  printf( "Chassis Type        : %s\n", chasstypes[k]);
   }
